feat(calculator): 소수 지원 및 특수 케이스 처리

+++ src/main/java/calculator/presentation/ApplicationRunner.java
@@ -16,7 +16,7 @@ public class ApplicationRunner {
 
     public void run() {
         String inputString = input.readExpression();
-        long result = calculator.calculate(inputString);
+        double result = calculator.calculate(inputString);
         output.printResult(result);
     }
